credit

+++ entities/dtos/requests/credit_request.go
@@ -2,6 +2,6 @@ package requests
 
 type CreditRequest struct {
 	ToUserID    uint64  `json:"to_user_id"`
-	Credit      float64 `json:"credit"`
+	Credit      float32 `json:"credit"`
 	Description string  `json:"description"`
 }
